feat/API15 Update plugin for API15
# Conflicts:
#	DalamudPlugin/.gitignore
#	DalamudPlugin/RaidsRewritten/Interop/ResourceLoader.cs
#	DalamudPlugin/RaidsRewritten/Memory/StatusCommonProcessor.cs
#	DalamudPlugin/RaidsRewritten/Memory/StatusCustomProcessor.cs
#	DalamudPlugin/RaidsRewritten/Memory/StatusFocusTargetProcessor.cs
#	DalamudPlugin/RaidsRewritten/Memory/StatusPartyListProcessor.cs
#	DalamudPlugin/RaidsRewritten/Memory/StatusProcessor.cs
#	DalamudPlugin/RaidsRewritten/Memory/StatusTargetInfoBuffDebuffProcessor.cs
#	DalamudPlugin/RaidsRewritten/Memory/StatusTargetInfoProcessor.cs
#	DalamudPlugin/RaidsRewritten/UI/View/MainWindow.DebugTab.cs

+++ DalamudPlugin/RaidsRewritten/EntityManager.cs
@@ -66,7 +66,7 @@ public sealed class EntityManager : IDalamudHook
         this.world.DeleteWith<Attack>();
     }
 
-    private void OnTerritoryChanged(ushort obj)
+    private void OnTerritoryChanged(uint obj)
     {
         ClearAllManagedEntities();
     }
